updated: HomeViewModel removed: MainViewModel.kt
