+++ app/src/main/java/dev/enesky/doodle/app/di/ViewModelModule.kt
@@ -1,7 +1,6 @@
 package dev.enesky.doodle.app.di
 
 import dev.enesky.core.domain.di.useCaseModule
-import dev.enesky.doodle.app.ui.main.MainViewModel
 import dev.enesky.feature.details.DetailsViewModel
 import dev.enesky.feature.login.LoginViewModel
 import dev.enesky.feature.main.HomeViewModel
@@ -19,7 +18,6 @@ val viewModelModule = lazyModule {
 
     includes(useCaseModule)
 
-    viewModelOf(::MainViewModel)
     viewModelOf(::LoginViewModel)
     viewModelOf(::HomeViewModel)
     viewModelOf(::DetailsViewModel)
